check for LID in db before hitting page
migrated the check in the database upwards in
process_links so that the function will strip the
key off the scraped link, check the db for it
and then skip it if it has already been logged,
rather than fetch the page, build the object
and then check for the key.  As a result
also moved the sleep time to the section of the code where
there is actually a page retrieved, this dramatically
speeds up the whole process of scraping new pages
as there is not a lot of redundant pinging of pages
that we have already hit up several times over
the course of the last week.  why did I do it that
way in the first place?1?!