adjusting to compensate for room sublets
The listings for units that are just a room are
missing most of the features that house or apartment
listings contain.  This was causing the various functions
to fail because the values were missing in the data structure.
There is still work to do here. But in testing a
listing object can be created and the various methods
can be called without causing failure...